// src/cinema/enums/cinema.enum.ts
export enum CinemaStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
  MAINTENANCE = 'maintenance',
}

export enum RoomStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
  MAINTENANCE = 'maintenance',
}

export enum SeatType {
  STANDARD = 'standard',
  VIP = 'vip',
  COUPLE = 'couple',
  WHEELCHAIR = 'wheelchair',
}